+++ src/Lykke.RabbitMqBroker/AssemblyDisplayName.cs
@@ -1,0 +1,16 @@
+using System.Reflection;
+
+namespace Lykke.RabbitMqBroker;
+
+internal abstract class AssemblyDisplayName
+{
+    protected AssemblyDisplayName(Assembly assembly)
+    {
+        var name = assembly.GetName();
+        Title = name.Name;
+        Version = name.Version?.ToString();
+    }
+    
+    public string Title { get; }
+    public string Version { get; }
+}