+++ src/Lykke.RabbitMqBroker/ConnectionAssemblyBasedDisplayName.cs
@@ -1,36 +1,22 @@
-using System.Reflection;
+using System.Text;
 
 namespace Lykke.RabbitMqBroker;
 
 /// <summary>
 /// Display name of the connection based on the assembly metadata (title and version)
+/// and library assembly metadata (title and version).
 /// </summary>
 internal abstract class ConnectionAssemblyBasedDisplayName
 {
-    private string _assemblyTitle;
-    private string _assemblyVersion;
+    private readonly HostAssemblyDisplayName _hostAssemblyDisplayName = new ();
+    private readonly LibraryAssemblyDisplayName _libraryAssemblyDisplayName = new ();
 
-    protected ConnectionAssemblyBasedDisplayName()
-    {
-        ExtractAssemblyMetadata();
-    }
-        
-    private void ExtractAssemblyMetadata()
-    {
-        var assemblyName = Assembly.GetEntryAssembly()?.GetName();
-            
-        _assemblyTitle = assemblyName?.Name;
-        _assemblyVersion = assemblyName?.Version?.ToString();
-    }
-        
     protected string BuildDisplayName(string prefix)
     {
-        var result = prefix;
-        if (!string.IsNullOrWhiteSpace(_assemblyTitle))
-            result += $" from {_assemblyTitle}";
-        if (!string.IsNullOrWhiteSpace(_assemblyVersion))
-            result += $" of version {_assemblyVersion}";
-            
-        return result;
+        var result = new StringBuilder(prefix);
+        result.Append($" from [{_hostAssemblyDisplayName}]");
+        result.Append($" using [{_libraryAssemblyDisplayName}]");
+
+        return result.ToString();
     }
 }