+++ src/Lykke.RabbitMqBroker/HostAssemblyDisplayName.cs
@@ -1,0 +1,15 @@
+using System.Reflection;
+
+namespace Lykke.RabbitMqBroker;
+
+internal class HostAssemblyDisplayName : AssemblyDisplayName
+{
+    public HostAssemblyDisplayName() : base(Assembly.GetEntryAssembly())
+    {
+    }
+
+    public override string ToString()
+    {
+        return $"{Title} {Version}";
+    }
+}